+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add changelog file (#13)
 - Log a "Done" message at the end of each successful operation (#14)
+- Option to get raw latest tag/version
 
 ## [0.1.0] - 2020-03-22